Separated OnlineLib Robot and used TextId for matching texts
