+++ app/src/androidTest/java/org/kiwix/kiwixmobile/main/TopLevelDestinationTest.kt
@@ -41,6 +41,7 @@ class TopLevelDestinationTest {
         }
       }
       openDrawer()
+
       clickBookmarksOnNavDrawer {
         clickOnTrashIcon()
         assertDeleteBookmarksDialogDisplayed()
